break 'utils' functionality into multiple modules
